+++ packages/jest-circus/src/__tests__/baseTest.test.ts
@@ -83,3 +83,72 @@ test('concurrent.each', () => {
 
   expect(stdout).toMatchSnapshot();
 });
+
+test('describe.skip + concurrent', () => {
+  const {stdout} = runTest(`
+    describe.skip('describe', () => {
+      beforeEach(() => {});
+      afterEach(() => { throw new Error('banana')});
+      test.concurrent('one', () => {
+        console.log('hello one');
+        throw new Error('kentucky')
+      });
+      test.concurrent('two', () => {
+        console.log('hello two');
+      });
+      test.concurrent('three', async () => {
+        console.log('hello three');
+        await Promise.resolve();
+      });
+    })
+  `);
+
+  expect(stdout).not.toEqual(expect.stringContaining('hello'));
+
+  expect(stdout).toMatchSnapshot();
+});
+
+test('describe.skip + concurrent.each', () => {
+  const {stdout} = runTest(`
+    describe.skip('describe', () => {
+      beforeEach(() => {});
+      afterEach(() => { throw new Error('banana')});
+      test.concurrent.each([
+        ['one'],
+        ['two'],
+        ['three'],
+      ])('%s', async (name) => {
+        console.log('hello %s', name);
+        await Promise.resolve();
+      });
+    })
+  `);
+
+  expect(stdout).not.toEqual(expect.stringContaining('hello'));
+
+  expect(stdout).toMatchSnapshot();
+});
+
+test('nested describe.skip + concurrent', () => {
+  const {stdout} = runTest(`
+    describe('describe', () => {
+      describe.skip('nested', () => {
+        test.concurrent('one', () => {
+          console.log('hello one');
+          throw new Error('kentucky')
+        });
+        test.concurrent('two', () => {
+          console.log('hello two');
+        });
+        test.concurrent('three', async () => {
+          console.log('hello three');
+          await Promise.resolve();
+        });
+      })
+    })
+  `);
+
+  expect(stdout).not.toEqual(expect.stringContaining('hello'));
+
+  expect(stdout).toMatchSnapshot();
+});